LF-4220 combine ProductDetails and ProductInfo
